Implement genome tree structure and utilities
This data structure is tree augmented with doubly-linked lists joining
certain branches. It can be searched recursively from root (genome) to
leaf (exon). Two levels, chromosome and gene, can also be iterated
through with LAST and NEXT links. This allows direct access to siblings.
It also permits arrays of pointers to nodes to be easily copied for
analysis, sorting, and whatnot outside the main data structure. Genes do
not link across chromosomes, which prevents false inferences of
adjacency.

This tree is loaded from a file of very specific format. This
specificity makes the program a bit brittle.