Implement test structure
I want to keep the tests close to the code. So I will put test_*
functions in each *.c file that I want to test. These functions will run
a battery of tests on the local functions, print the results, and return
TRUE (or technically, 1) if any of the test fail.

A test.c file includes all files that contain test functions. It manages
global tests as well as the unit tests.

I have wired a test subcommand into the UI. If the 'test' flag is chosen
by the user, main will call the test_all() function defined in test.c
and exit. The output will be a stream of test results and an exit status
of 0 if all tests pass and 1 if any fail.