Fix sort bug
Finding the flanking syntenic regions around a query require finding the
nearest interval STOP on the left and START on the right. Each Block
stores a stopid, which is suppose to be an index in an array of Block
pointers that is sorted by STOP position. However, I was not actually
building this array, so these indices ended up being random locations
in memory.