Implement flank return in itree
Previously itree returned only the nearest interval on one side of a
non-overlapping query. Finding the other flank was left as a problem for
the user. But unfortunately, this task is not trivial and is best done
within the context of the interval tree.

Now if no overlap is found, the nearest flanking intervals are returned.
No handling of ties is done.

Furthermore, I have added two additional flags to the IntervalResult
structure that is returned by the get_overlap functions:

 1. leftmost - set when query ends before any interval begins
 2. rightmost - set when query begins after all intervals end

I have not yet integrated these changes with the rest of the program;
this commit only runs the interval tree and prints results.