Sort linkids relative to query start positions
Previously linkids were handled in the shellscript. I've now moved it to
the C-side load_synmap function. This is slower, since it requires an
extra pass through the synteny map, but it ensures the query intervals
and linkid are in the same order. This is required in the
populate_contiguous_map function.