Remove offsets
I was previously making inbetween search intervals exclude the start
stop positions of the flanks. For example, if a search space is between
the intervals (100,200) and (300,400), I would set the search interval
to (201,299). However, this results in negative lengths for mappings
between immediately adjacent intervals. For example, (100,200) and
(201,300). By offsettin, I get a search interval of (200 + 1, 201 - 1)
and thus a length of -1.

So I have removed the offsets. Now a hit inside and outside an interval
have search intervals that overlap by one, which I don't like. But it is
better than negative lengths.

I have also updated all test cases to reflect these changes.