Add scaffold length handling
If a search interval extends beyond the furthest syntenic interval, it
should snap to the end of the scaffold. However, this requires knowledge
of the lengths of each scaffold. Previously, I did not give this
information to Synder, so I just assigned a dummy value of 999999999 to
the right extremists.

In this commit, I add handling of scaffold length input files (one for
query and one for target). If these files are given, Synder will store
scaffold lengths in the Contig structures, and snap to the scaffold
endpoint when needed. Otherwise, it uses the same dummy value as before.