Add start/stop and in/out base offsets
Initially I assumed all input is 0-based and I wrote only to 0-base.
However, this requires the used to do extra work, so I added a 1-base
option.

Unfortunately, 0- and 1-base options are not the only possibilities.
There are also programs that output starts in 0-base and stops in
1-base. So I let the user to input a vector of offsets. For example,
"0111" takes 0-based start input and 1-based stop input; then it writes
1-based start and 1-based stop output.