Major refactor
Memory for all Block structures in a Contig is allocated once when the
Contig is initialized.

Blocks are now a full blown iterable data structure. Each Block contains
nine links to other blocks (see the header for details). Rather than
for-looping across an array of Blocks, we can now navigate completely by
linking. This allows deletion of nodes, which is needed for the next
iteration of the algorithm.

The current code compiles without error. But I haven't run the tests
yet. Many bug fixes will follow.