k = 0 correct 3rd generation prototype
When k, the number of interrupting intervals allowed in a contiguous
set, is 0, then the 3rd and 2nd generation algorithms should have
identical output.

I have modified link_contiguous_blocks to what I think is roughly
generation 3. But it breaks on any real tests for k > 0. For k=0,
however, it is correct (it passes all memory and logic test and yields
identical results for the full Brassicaceae dataset).

Now I just need to correct it for k > 0 ...