Implement leaky version of algorithm 2.3
Currently calling free_Synmap results in a segfault. So I just leak that
memory. This doesn't really matter, since it is a terminal free, but it
is not elegant. The data leak is caused by double-freeing of the
ContiguousSet pointers I give to itree. I'll fix the root problem in the
next big refactor.

This commit allows skipping of k intervals separating two elements in a
contiguous set, if all k intervals have one side in a scaffold not in
the contiguous set.

The following needs to be done for algorithm 3.0:
 * merge overlapping elements
 * test query scores
 * implement target scores
 * implement composite scores
 * write all three scores to output (?)
 * report number of elements skipped