Add setid to search output and block dump option
The contiguous set ids are written to search output and to the block
dump output. The latter is equivalent to the synteny map input with an
additional contiguous set id as the last column. Since contiguous set
id assignment is deterministic, the ids can be used to associate search
intervals with their parent contiguous sets. This will be useful in
downstream statists and diagnostics.