Stubs compile
I written myself into a highly broken state. In this commit I've managed
to get the program to compile by commenting out most of the function
blocks in source files. The overall template and inheritance structure
is now basically valid (it compiles).

The next step is to incrementally uncomment the function blocks and
modify as needed.