Tag deleted blocks using NULL over field
Previously when blocks were deleted (when removing overlapping blocks) I
memset the structures to 0. But that isn't great practice, especially in
C++. Also it leads to undefined behaviour, e.g. in the float member.
Finally, it is unnecessary and slow.

So instead I will simply assign the `over` field to NULL.