Improve score merging and tests
When doubly-overlapping blocks are merged the scores also need to be
merged. Previously I used the following equation:

 da (la - lo) + db (lb - lo) + lo (da + db) / 2            E1

Where
 - *da* and *db* are the score densities of blocks *a* and *b*
 - *la*, *lb* and *lo* are the lengths of *a*, *b*, and their overlap

E1 is problematic for two reasons. First, it doesn't really make sense
to average the overlap scores. Second, iterative pairwise averaging is
assymetric, giving higher weight to the blocks merged later.

I have replaced this approach with taking the max of the overlap scores:

 da (la - lo) + db (lb - lo) + lo * max(da, db)            E2

This fixes the first problem.

The second problem is a bit trickier, since it would require tracking
sub-intervals. It would be a pain to implement and probably would have
little effect on any real dataset. So I am content with an imperfect
solution for now.

Minutia:
    - Added tests for the overlap cases as scored by E2.
    - Added debug print statements when merging
    - Added dedicated function for testing `synder dump` output