Fix bug in contig corner calculations
The core data structure of Synder is vast and interconnected. The system
passes through many transitional states where the invariants are not
respected. My functions are impure or depend on the non-obvious
implementation details of other scopres. Ownership is spread between
many entities. The impurity and high-dimensionallity of my
implementation complicates reasoning about its correctness. So I rely on
tests. But the tests are fragile, as is the test script itself.