Read %s rather than %c in gff strand column
Previously I treated the strand column as a character column (which it
should be in a properly formatted GFF file), but this breaks when I feed
in GFF data tables (as loaded by `synder::read_gff`) which replace
unknown strand with "NA"; 'N' maps to column 7 and 'A' to column 8. I
fixed this by treating the strand column as a string (%s) instead of a
character. This is not as a tight as my former solution, and ultimately
I should replace sscanf with a better parser, but it works for now.