Set default offsets to 000000
Previously I had the default offset set to 010011, which is the
configuration that works in the Satsuma -> GFF -> bioconductor pipeline.
However, a more natural default is to 0- or 1-base everything. I choose
to 0-base because that requires no changes to my tests. But 1-base would
also be reasonable, perhaps more so, since this is the R convention.