Fixes #2 where read failed for long 9th column
I have replaced all C-style IO functions with their C++ iostream
counterparts. String fields can now be of arbitrary length, which fixes
issue #2.

I delimit based on whitespace, rather than tabs, even in the
tab-delimited files. This will not normally be a problem, since there
should not be any non-TAB whitespace in a GFF file (or any of the other
input files), however, well, people sin.

I still do not thoroughly type-check the inputs. To remedy this, I
should write a battery of format-tests and then recode till they all
pass. But this is a problem for another day.