Add cast functions
These are functions such as `as_synmap` that convert something that can
be cast as a data.frame to the appropriate synder type. The `as_*`
family does full checking. The `read_*` family is now uses `as_*` for
checking.

I ought to add handling for converting to GFF from a GRanges
bioconductor object.