Standardize column names
Name changes in all tables:

 * con     -> seqid
 * qcon    -> qseqid
 * tcon    -> tseqid
 * seqname -> attr

The last change is because `synder` uses the 9th GFF column as the
interval name. However, it really isn't proper to refer to this column
in general as the seqname, since, 1) it isn't the name of a sequence,
but rather an interval, and 2) it isn't name at all, but rather a list
of attributes.