Handle swapping of contig length files C-side
Maddeningly, everything broke when I recompiled the source. The load
order for the packages was off, or something. I cursed and banged on the
code until it suddenly worked again. R is killing me. Doubtless there is
a documentation page somewhere that explains the byzantine rules, but
there is no excuse for this sort of instability. R is simply a poorly
designed language.