Unify missing contig warnings
Previously there was one warnig emitted for every contig in the GFF that
was missing in the searched synteny map. Now there is a single
conglomerate warning message that specifies how man contigs are missing
(and states the total number) and reports the missing contigs in a list.

There is a trailing comma in the list of missing contigs. This is ugly,
but doesn't really matter.