Simplify offsets
Now only the offsets for the synteny map can be set by the user. The GFF
is now assumed to be 1-based. The offsets for the output were made
irrelevant when I refactored the output as Bioconductor objects (which
are 1-based).

Since I now assume the GFF is 1-based, I need to add a check intervals
for `start > 0`. Of course, I should check the synteny maps as well for
the same problem.

Also, I removed the test for offset setting. I need to add a new one to
check 1) the correct handling of synteny maps and 2) the catching of
offset-invalid files.