Pass only synteny map offsets C-side
I still use 4 offsets internally, but the GFF offsets are hard-coded.
Only 2 of the 6 offsets are passed in, but this could easily be changed
(by updating the Synmap constructor, which reads the offsets passed from
R).

I still have to convert to 1-base on the R-side. I think this is a
little awkward, so if it is convenient, I'll move it C-side.