Write flag_summary function
This was adapted from Fagin. I should rethink the approach, though.
Specifically, unbound is a bit fishy, since it does not distinguish
between cases where a link is between the unbound edges.

It might be better if the boolean was replaced with count, specifying
how many elements fall into each category.

Also, the `unassembled` and `scrambled` terms could be added to the
summaries.

Finally, I need to add some tests.

For the interpreters in general, I should consult the literature to find
good metrics for describing synteny.